set float32 to avoid casting errors

+++ racecar_gym/bullet/sensors.py
@@ -116,9 +116,9 @@ class Lidar(BulletSensor[NDArray[Any, Float]]):
         results = p.rayTestBatch(self._from, self._to, 0,
             parentObjectUniqueId=self.body_id,
             parentLinkIndex=self.joint_index)
-        hit_fractions = np.array(results, dtype=object)[:, 2].astype(dtype=float)
+        hit_fractions = np.array(results, dtype=object)[:, 2].astype(dtype=np.float32)
         ranges = self._config.range * hit_fractions + self._config.min_range
-        noise = np.random.uniform(1.0 - self._config.accuracy, 1.0 + self._config.accuracy, size=ranges.shape)
+        noise = np.random.uniform(1.0 - self._config.accuracy, 1.0 + self._config.accuracy, size=ranges.shape).astype(np.float32)
         scan = np.clip(ranges * noise, a_min=self._config.min_range, a_max=self._config.range)
 
         if self._config.debug:
